+++ Exposure/Entitlements/Request/PlayCatchup.swift
@@ -17,7 +17,7 @@ import Foundation
 /// * `format` HLS
 ///
 /// If there is no current program live will return `NOT_ENABLED`.
-public struct PlayCatchup: Exposure, DRMRequest {
+public struct PlayCatchup: ExposureType, DRMRequest {
     public typealias Response = PlaybackEntitlement
     
     /// Channel of the program to play.