+++ CHANGELOG.md
@@ -17,6 +17,7 @@ NO RELEASE DATE SET YET
 * `EMP-10270` `Airing` is now distinct from fetching `Program`
 * `EMP-10322` `Image.Orientation` made `Equatable`.
 * Integrated `Download` entitlement requests and validation.
+* Renamed `Exposure` protocol to `ExposureType`
 
 #### Bug fixes
 * `EMP-10269` `FetchEPG` now respects the *onlyPublished* parameter. Fixed url encoding.